+++ src/SFML/Window/Joystick.cpp
@@ -48,7 +48,7 @@ unsigned int Joystick::getButtonCount(unsigned int joystick)
 ////////////////////////////////////////////////////////////
 bool Joystick::hasAxis(unsigned int joystick, Axis axis)
 {
-    return priv::JoystickManager::getInstance().getCapabilities(joystick).axes[axis];
+    return priv::JoystickManager::getInstance().getCapabilities(joystick).axes[static_cast<size_t>(axis)];
 }
 
 
@@ -62,7 +62,7 @@ bool Joystick::isButtonPressed(unsigned int joystick, unsigned int button)
 ////////////////////////////////////////////////////////////
 float Joystick::getAxisPosition(unsigned int joystick, Axis axis)
 {
-    return priv::JoystickManager::getInstance().getState(joystick).axes[axis];
+    return priv::JoystickManager::getInstance().getState(joystick).axes[static_cast<size_t>(axis)];
 }
